Finish SetUp

+++ GStar/SimpleExample.cpp
@@ -4,7 +4,7 @@ GStar::TextureData * SimpleExample::getdata()
 {
 	int nx = 200;
 	int ny = 100;
-	float data[200*100* 4];
+	uint8_t data[200*100* 4];
 	for (int i = 0; i < 200; i++) {
 		for (int j = 0; j < 100; j++) {
 			float r = float(i) / float(nx);
